+++ monkey/tests/monkey_island/cc/environment/test_user_creds.py
@@ -1,37 +1,45 @@
-from unittest import TestCase
-
 from monkey_island.cc.environment.user_creds import UserCreds
 
+TEST_USER = "Test"
+TEST_HASH = "abc1231234"
+TEST_SALT = b"$2b$12$JA7GdT1iyfIsquF2cTZv2."
 
-class TestUserCreds(TestCase):
-    def test_to_dict(self):
-        user_creds = UserCreds()
-        self.assertDictEqual(user_creds.to_dict(), {})
 
-        user_creds = UserCreds(username="Test")
-        self.assertDictEqual(user_creds.to_dict(), {"user": "Test"})
+def test_bool_true():
+    assert UserCreds(TEST_USER, TEST_HASH)
 
-        user_creds = UserCreds(password_hash="abc1231234")
-        self.assertDictEqual(user_creds.to_dict(), {"password_hash": "abc1231234"})
 
-        user_creds = UserCreds(username="Test", password_hash="abc1231234")
-        self.assertDictEqual(user_creds.to_dict(), {"user": "Test", "password_hash": "abc1231234"})
+def test_bool_false_empty_password_hash():
+    assert not UserCreds(TEST_USER, "")
 
-    def test_to_auth_user(self):
-        user_creds = UserCreds(username="Test", password_hash="abc1231234")
-        auth_user = user_creds.to_auth_user()
-        self.assertEqual(auth_user.id, 1)
-        self.assertEqual(auth_user.username, "Test")
-        self.assertEqual(auth_user.secret, "abc1231234")
 
-        user_creds = UserCreds(username="Test")
-        auth_user = user_creds.to_auth_user()
-        self.assertEqual(auth_user.id, 1)
-        self.assertEqual(auth_user.username, "Test")
-        self.assertEqual(auth_user.secret, "")
+def test_bool_false_empty_user():
+    assert not UserCreds("", TEST_HASH)
 
-        user_creds = UserCreds(password_hash="abc1231234")
-        auth_user = user_creds.to_auth_user()
-        self.assertEqual(auth_user.id, 1)
-        self.assertEqual(auth_user.username, "")
-        self.assertEqual(auth_user.secret, "abc1231234")
+
+def test_bool_false_empty_user_and_password_hash():
+    assert not UserCreds("", "")
+
+
+def test_to_dict_empty_creds():
+    user_creds = UserCreds("", "")
+    assert user_creds.to_dict() == {}
+
+
+def test_to_dict_full_creds():
+    user_creds = UserCreds(TEST_USER, TEST_HASH)
+    assert user_creds.to_dict() == {"user": TEST_USER, "password_hash": TEST_HASH}
+
+
+def test_to_auth_user_full_credentials():
+    user_creds = UserCreds(TEST_USER, TEST_HASH)
+    auth_user = user_creds.to_auth_user()
+    assert auth_user.id == 1
+    assert auth_user.username == TEST_USER
+    assert auth_user.secret == TEST_HASH
+
+
+def test_member_values(monkeypatch):
+    creds = UserCreds(TEST_USER, TEST_HASH)
+    assert creds.username == TEST_USER
+    assert creds.password_hash == TEST_HASH